+++ plugins/misc.py
@@ -25,14 +25,13 @@
 import os
 from datetime import datetime
 
+import cfscrape
 from bs4 import BeautifulSoup as bs
 from htmlwebshot import WebShot
 from img2html.converter import Img2HTMLConverter
-from requests import get
 
 from . import (
     async_searcher,
-    eor,
     fast_download,
     get_random_user_data,
     get_string,
@@ -53,17 +52,17 @@ def gib_link(link):
 @ultroid_cmd(pattern="eod ?(.*)")
 async def diela(e):
     match = e.pattern_match.group(1)
-    m = await eor(e, get_string("com_1"))
+    m = await e.eor(get_string("com_1"))
     li = "https://daysoftheyear.com"
     te = "🎊 **Events of the Day**\n\n"
     if match:
         date = match.split("/")[0]
         month = match.split("/")[1]
-        li += "/days/2021/" + month + "/" + date
+        li += "/days/2021-2022/" + month + "/" + date
         te = get_string("eod_2").format(match)
     else:
         da = datetime.today().strftime("%F").split("-")
-        li += "/days/2021/" + da[1] + "/" + da[2]
+        li += "/days/2021-2022/" + da[1] + "/" + da[2]
     ct = requests.get(li).content
     bt = bs(ct, "html.parser", from_encoding="utf-8")
     ml = bt.find_all("a", "js-link-target", href=re.compile("daysoftheyear.com/days"))
@@ -77,14 +76,16 @@ async def diela(e):
 )
 async def pinterest(e):
     m = e.pattern_match.group(1)
-    get_link = get(gib_link(m)).text
-    hehe = bs(get_link, "html.parser")
+    if not m:
+        return await e.eor("`Give pinterest link.`", time=3)
+    scrape = cfscrape.create_scraper()
+    hehe = bs(scrape.get(gib_link(m)).text, "html.parser")
     hulu = hehe.find_all("a", {"class": "download_button"})
     if len(hulu) < 1:
-        await eor(e, "`Wrong link or private pin.`", time=5)
+        await e.eor("`Wrong link or private pin.`", time=5)
     elif len(hulu) > 1:
-        video = await fast_download(hulu[0]["href"])
-        thumb = await fast_download(hulu[1]["href"])
+        video, _ = await fast_download(hulu[0]["href"])
+        thumb, _ = await fast_download(hulu[1]["href"])
         await e.delete()
         await e.client.send_file(e.chat_id, video, thumb=thumb, caption=f"Pin:- {m}")
         [os.remove(file) for file in [video, thumb]]
@@ -97,15 +98,15 @@ async def pinterest(e):
 async def mobs(e):
     mat = e.pattern_match.group(1)
     if not mat:
-        await eor(e, "Please Give a Mobile Name to look for.")
+        await e.eor("Please Give a Mobile Name to look for.")
     query = mat.replace(" ", "%20")
     jwala = f"https://gadgets.ndtv.com/search?searchtext={query}"
     c = await async_searcher(jwala)
     b = bs(c, "html.parser", from_encoding="utf-8")
     co = b.find_all("div", "rvw-imgbox")
     if not co:
-        return await eor(e, "No Results Found!")
-    bt = await eor(e, get_string("com_1"))
+        return await e.eor("No Results Found!")
+    bt = await e.eor(get_string("com_1"))
     out = "**📱 Mobile / Gadgets Search**\n\n"
     li = co[0].find("a")
     imu, title = None, li.find("img")["title"]
@@ -122,13 +123,15 @@ async def mobs(e):
         ty = fp.findNext()
         out += f"- **{ty.text}** - `{ty.findNext().text}`\n"
     out += "_"
+    if imu == []:
+        imu = None
     await e.reply(out, file=imu)
     await bt.delete()
 
 
 @ultroid_cmd(pattern="randomuser")
 async def _gen_data(event):
-    x = await eor(event, get_string("com_1"))
+    x = await event.eor(get_string("com_1"))
     msg, pic = await get_random_user_data()
     await event.reply(file=pic, message=msg)
     await x.delete()
@@ -139,8 +142,8 @@ async def _gen_data(event):
 )
 async def _(e):
     if not e.reply_to_msg_id:
-        return await eor(e, get_string("ascii_1"))
-    m = await eor(e, get_string("ascii_2"))
+        return await e.eor(get_string("ascii_1"))
+    m = await e.eor(get_string("ascii_2"))
     img = await (await e.get_reply_message()).download_media()
     char = "■" if not e.pattern_match.group(1) else e.pattern_match.group(1)
     converter = Img2HTMLConverter(char=char)
